+++ framework/src/main/java/org/checkerframework/common/value/ValueVisitor.java
@@ -6,7 +6,11 @@ import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.Tree;
 import com.sun.source.tree.TypeCastTree;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Objects;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 import javax.lang.model.element.AnnotationMirror;
@@ -22,6 +26,7 @@ import org.checkerframework.common.basetype.BaseTypeVisitor;
 import org.checkerframework.common.value.qual.IntRangeFromGTENegativeOne;
 import org.checkerframework.common.value.qual.IntRangeFromNonNegative;
 import org.checkerframework.common.value.qual.IntRangeFromPositive;
+import org.checkerframework.common.value.qual.IntVal;
 import org.checkerframework.common.value.qual.StaticallyExecutable;
 import org.checkerframework.common.value.util.NumberUtils;
 import org.checkerframework.common.value.util.Range;
@@ -29,10 +34,13 @@ import org.checkerframework.dataflow.qual.Pure;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedDeclaredType;
 import org.checkerframework.framework.type.visitor.AnnotatedTypeScanner;
+import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.AnnotationUtils;
 import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.TreeUtils;
+import org.checkerframework.javacutil.TypeKindUtils;
 import org.checkerframework.javacutil.TypesUtils;
+import org.plumelib.util.CollectionsPlume;
 
 /** Visitor for the Constant Value type system. */
 public class ValueVisitor extends BaseTypeVisitor<ValueAnnotatedTypeFactory> {
@@ -288,6 +296,7 @@ public class ValueVisitor extends BaseTypeVisitor<ValueAnnotatedTypeFactory> {
         && exprAnno != null
         && atypeFactory.isIntRange(castAnno)
         && atypeFactory.isIntRange(exprAnno)) {
+      // TODO: Should handle signedness.
       final Range castRange = atypeFactory.getRange(castAnno);
       final TypeKind castTypeKind = castType.getKind();
       if (castTypeKind == TypeKind.BYTE && castRange.isByteEverything()) {
@@ -321,7 +330,108 @@ public class ValueVisitor extends BaseTypeVisitor<ValueAnnotatedTypeFactory> {
         }
       }
     }
+
     return super.visitTypeCast(tree, p);
+  }
+
+  // Problem:  At this point, types are like: (@IntVal(-1) byte, @IntVal(255) int) and knowledge of
+  // signedness is gone.  But, I can use castType's underlying type to infer correctness.
+  @Override
+  protected boolean isTypeCastSafe(AnnotatedTypeMirror castType, AnnotatedTypeMirror exprType) {
+    TypeKind castTypeKind = TypeKindUtils.primitiveOrBoxedToTypeKind(castType.getUnderlyingType());
+    TypeKind exprTypeKind = TypeKindUtils.primitiveOrBoxedToTypeKind(exprType.getUnderlyingType());
+    if (TypeKindUtils.isIntegral(castTypeKind) && TypeKindUtils.isIntegral(exprTypeKind)) {
+      AnnotationMirrorSet castAnnos = castType.getAnnotations();
+      AnnotationMirrorSet exprAnnos = exprType.getAnnotations();
+      if (castAnnos.equals(exprAnnos)) {
+        return true;
+      }
+      assert castAnnos.size() == 1;
+      assert exprAnnos.size() == 1;
+      AnnotationMirror castAnno = castAnnos.first();
+      AnnotationMirror exprAnno = exprAnnos.first();
+      boolean castAnnoIsIntVal = atypeFactory.areSameByClass(castAnno, IntVal.class);
+      boolean exprAnnoIsIntVal = atypeFactory.areSameByClass(exprAnno, IntVal.class);
+      if (castAnnoIsIntVal && exprAnnoIsIntVal) {
+        List<Long> castValues = atypeFactory.getIntValues(castAnno);
+        List<Long> exprValues = atypeFactory.getIntValues(exprAnno);
+        if (castValues.size() == 1 && exprValues.size() == 1) {
+          // Special-case singleton sets for speed.
+          switch (castTypeKind) {
+            case BYTE:
+              return castValues.get(0).byteValue() == exprValues.get(0).byteValue();
+            case INT:
+              return castValues.get(0).intValue() == exprValues.get(0).intValue();
+            case SHORT:
+              return castValues.get(0).shortValue() == exprValues.get(0).shortValue();
+            default:
+              return castValues.get(0).longValue() == exprValues.get(0).longValue();
+          }
+        } else {
+          switch (castTypeKind) {
+            case BYTE:
+              {
+                TreeSet<Byte> castValuesTree =
+                    new TreeSet<Byte>(CollectionsPlume.mapList(Number::byteValue, castValues));
+                TreeSet<Byte> exprValuesTree =
+                    new TreeSet<Byte>(CollectionsPlume.mapList(Number::byteValue, exprValues));
+                return sortedSetEquals(castValuesTree, exprValuesTree);
+              }
+            case INT:
+              {
+                TreeSet<Integer> castValuesTree =
+                    new TreeSet<Integer>(CollectionsPlume.mapList(Number::intValue, castValues));
+                TreeSet<Integer> exprValuesTree =
+                    new TreeSet<Integer>(CollectionsPlume.mapList(Number::intValue, exprValues));
+                return sortedSetEquals(castValuesTree, exprValuesTree);
+              }
+            case SHORT:
+              {
+                TreeSet<Short> castValuesTree =
+                    new TreeSet<Short>(CollectionsPlume.mapList(Number::shortValue, castValues));
+                TreeSet<Short> exprValuesTree =
+                    new TreeSet<Short>(CollectionsPlume.mapList(Number::shortValue, exprValues));
+                return sortedSetEquals(castValuesTree, exprValuesTree);
+              }
+            default:
+              {
+                TreeSet<Long> castValuesTree = new TreeSet<>(castValues);
+                TreeSet<Long> exprValuesTree = new TreeSet<>(exprValues);
+                return sortedSetEquals(castValuesTree, exprValuesTree);
+              }
+          }
+        }
+      }
+    }
+
+    return super.isTypeCastSafe(castType, exprType);
+  }
+
+  // TODO: After plume-util 1.6.6 is released, use this method from CollectionsPlume instead.
+  /**
+   * Returns true if the two sets contain the same elements in the same order.
+   *
+   * @param <T> the type of elements in the sets
+   * @param set1 the first set to compare
+   * @param set2 the first set to compare
+   * @return true if the two sets contain the same elements in the same order
+   */
+  private static <T> boolean sortedSetEquals(SortedSet<T> set1, SortedSet<T> set2) {
+    @SuppressWarnings("interning:not.interned")
+    boolean sameObject = set1 == set2;
+    if (sameObject) {
+      return true;
+    }
+    if (set1.size() != set2.size()) {
+      return false;
+    }
+    for (Iterator<T> itor1 = set1.iterator(), itor2 = set2.iterator(); itor1.hasNext(); ) {
+      boolean same = Objects.equals(itor1.next(), itor2.next());
+      if (!same) {
+        return false;
+      }
+    }
+    return true;
   }
 
   /**