+++ framework/src/main/java/org/checkerframework/common/value/util/NumberUtils.java
@@ -9,9 +9,19 @@ import org.plumelib.util.CollectionsPlume;
 /** Utility routines for manipulating numbers. */
 public class NumberUtils {
 
-  /** Converts a {@code List<A>} to a {@code List<B>}, where A and B are numeric types. */
+  /**
+   * Converts a {@code List<A>} to a {@code List<B>}, where A and B are numeric types. This method
+   * is ignorant of signedness.
+   */
   public static List<? extends Number> castNumbers(
       TypeMirror type, List<? extends Number> numbers) {
+    return castNumbers(type, false, numbers);
+  }
+
+  /** Converts a {@code List<A>} to a {@code List<B>}, where A and B are numeric types. */
+  @SuppressWarnings("unchecked")
+  public static List<? extends Number> castNumbers(
+      TypeMirror type, boolean isUnsigned, List<? extends Number> numbers) {
     if (numbers == null) {
       return null;
     }
@@ -21,7 +31,12 @@ public class NumberUtils {
     }
     switch (typeKind) {
       case BYTE:
-        return CollectionsPlume.mapList(Number::byteValue, numbers);
+        if (isUnsigned) {
+          return CollectionsPlume.<Number, Short>mapList(
+              NumberUtils::byteValueUnsigned, (Iterable<Number>) numbers);
+        } else {
+          return CollectionsPlume.mapList(Number::byteValue, numbers);
+        }
       case CHAR:
         return CollectionsPlume.mapList(Number::intValue, numbers);
       case DOUBLE:
@@ -39,6 +54,14 @@ public class NumberUtils {
     }
   }
 
+  static Short byteValueUnsigned(Number n) {
+    short result = n.byteValue();
+    if (result > 127) {
+      result = (short) (result - 256);
+    }
+    return result;
+  }
+
   /**
    * Return a range that restricts the given range to the given type. That is, return the range
    * resulting from casting a value with the given range.
@@ -48,13 +71,25 @@ public class NumberUtils {
    * @return the intersection of the given range and the possible values of the given type
    */
   public static Range castRange(TypeMirror type, Range range) {
+    return castRange(type, false, range);
+  }
+
+  /**
+   * Return a range that restricts the given range to the given type. That is, return the range
+   * resulting from casting a value with the given range.
+   *
+   * @param type the type for the cast; the result will be within it
+   * @param range the original range; the result will be within it
+   * @return the intersection of the given range and the possible values of the given type
+   */
+  public static Range castRange(TypeMirror type, boolean isUnsigned, Range range) {
     TypeKind typeKind = TypeKindUtils.primitiveOrBoxedToTypeKind(type);
     if (typeKind == null) {
       throw new UnsupportedOperationException(type.toString());
     }
     switch (typeKind) {
       case BYTE:
-        return range.byteRange();
+        return range.byteRange(isUnsigned);
       case CHAR:
         return range.charRange();
       case SHORT: