+++ framework/src/main/java/org/checkerframework/common/value/util/Range.java
@@ -402,6 +402,23 @@ public class Range {
    * overflow rules, e.g., Byte.MAX_VALUE + 1 is converted to Byte.MIN_VALUE.
    */
   public Range byteRange() {
+    return byteRange(false);
+  }
+
+  /**
+   * Converts this range to a 8-bit byte range.
+   *
+   * <p>If {@link #ignoreOverflow} is true and one of the bounds is outside the Byte range, then
+   * that bound is set to the bound of the Byte range.
+   *
+   * <p>If {@link #ignoreOverflow} is false and this range is too wide, i.e., wider than the full
+   * range of the Byte class, return BYTE_EVERYTHING.
+   *
+   * <p>If {@link #ignoreOverflow} is false and the bounds of this range are not representable as
+   * 8-bit integers, convert the bounds to Byte type in accordance with Java twos-complement
+   * overflow rules, e.g., Byte.MAX_VALUE + 1 is converted to Byte.MIN_VALUE.
+   */
+  public Range byteRange(boolean isUnsigned) {
     if (this.isNothing()) {
       return this;
     }